Use size_t and uint64_t instead of long long in ed25519/.

+++ core/crypto/ed25519/ed25519_ref10.h
@@ -8,13 +8,13 @@ int ed25519_ref10_pubkey(unsigned char* pk, const unsigned char* sk);
 
 int ed25519_ref10_open(
     const unsigned char* sig,
-    const unsigned char* m, unsigned long long mlen,
+    const unsigned char* m, size_t mlen,
     const unsigned char*pk
 );
 
 int ed25519_ref10_sign(
     unsigned char* sig,
-    const unsigned char* m, unsigned long long mlen,
+    const unsigned char* m, size_t mlen,
     const unsigned char* sk, const unsigned char* pk
 );
