+++ core/crypto/ed25519/sign.cpp
@@ -6,7 +6,7 @@
 
 int crypto_sign(
   unsigned char *sig,
-  const unsigned char* m, unsigned long long mlen,
+  const unsigned char* m, size_t mlen,
   const unsigned char* sk, const unsigned char* pk
 )
 {
